requirements.txt
disnake>=2.9.1
aiohttp>=3.9.3
pytesseract>=0.3.10
pillow>=10.2.0
python-dotenv>=1.0.1